move the logic of the airlock request files section to its own file
